Refactoring matchers. Base class w/composed searchers.
