+++ pyproject.toml
@@ -24,21 +24,21 @@ pytest = "^5.4.3"
 coverage = {version = "^5.1", extras = ["toml"]}
 pytest-cov = "^2.10.0"
 pytest-mock = "^3.1.1"
-flake8-black = "^0.2"
+flake8-black = "0.*"
 flake8-bugbear = "^20.1.4"
 flake8-bandit = "^2.1.2"
 safety = "^1.9.0"
 flake8-import-order = "^0.18"
-mypy = "^0.782"
+mypy = "0.*"
 typeguard = "^2.9.1"
 flake8-annotations = "^2.1.0"
 flake8-docstrings = "^1.5.0"
 darglint = "^1.4.1"
-xdoctest = "^0.12"
+xdoctest = "0.*"
 sphinx = "^3.1.1"
 rstcheck = "^3.3.1"
 sphinx-autodoc-typehints = "^1.11.0"
-sphinx-autobuild = "^0.7"
+sphinx-autobuild = "0.*"
 codecov = "^2.1.7"
 
 [tool.poetry.plugins."spacy_factories"]
